MoviesDataService class to control the file generation
