+++ step2/src/Step_Two_Filter_Movies/RUNME.md
@@ -15,6 +15,12 @@ On Step2/src/Step_Two_Filter_Movies folder execute:
 ```
 mvn spring-boot:run "-Dspring-boot.run.arguments=--spring.profiles.active=dev --decade=2010 --output=2010s-movies.json"
 ```
+Alternatively, call the jar created in the target folder: 
+
+```
+java -jar -Dspring.profiles.active=dev ./target/Step_Two_Filter_Movies-1.0.0.jar "--decade=2000" "--output=2000s-movies.json"
+```
+
 
 The two minus signs should be used as described   
 
@@ -31,3 +37,4 @@ The two minus signs should be used as described
 > should be lowercase   
 > the file will have same name informed.   
 > by default, the file will be save on same folder that is movides.json, data folder.   
+> If the same file name exits, the validation will stop the process and an exception will be thrown.
